move find into router to access isExternal
