+++ src/services/routeMatchRules.ts
@@ -6,13 +6,6 @@ export const isNamedRoute: RouteMatchRule = (route) => {
   return 'name' in route.matched && !!route.matched.name
 }
 
-export const routeHostMatches: RouteMatchRule = (route, url) => {
-  const { host: urlHost } = createMaybeRelativeUrl(url)
-  const { host: routeHost } = createMaybeRelativeUrl(route.path.toString())
-
-  return urlHost === routeHost
-}
-
 export const routePathMatches: RouteMatchRule = (route, url) => {
   const { pathname } = createMaybeRelativeUrl(url)
   const pathPattern = generateRoutePathRegexPattern(route)
